HackRF Portapack test files.
You can add these to your bleTX directory on the SD card of your HackRF portapack in order to test any device meant to alert the presence of a Flock Security camera. I took the databases provided by ColonelPanic in the creation of his detection firmware for the seeedstudio esp32S3 and made fake bluetooth packets to send using the Hackrf for testing purposes.